+++ sampleapp/src/main/java/com/bignerdranch/expandablerecyclerviewsample/MainActivity.java
@@ -3,12 +3,9 @@ package com.bignerdranch.expandablerecyclerviewsample;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
 
-import com.bignerdranch.expandablerecyclerviewsample.R;
-import com.bignerdranch.expandablerecyclerviewsample.linear.horizontal.HorizontalLinearRecyclerViewSampleActivity;
 import com.bignerdranch.expandablerecyclerviewsample.linear.vertical.VerticalLinearRecyclerViewSampleActivity;
 
 /**
@@ -21,9 +18,6 @@ import com.bignerdranch.expandablerecyclerviewsample.linear.vertical.VerticalLin
 public class MainActivity extends AppCompatActivity {
 
     private Button mVerticalSampleButton;
-    private Button mHorizontalSampleButton;
-    private Button mGridSampleButton;
-    private Toolbar mToolbar;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,10 +26,6 @@ public class MainActivity extends AppCompatActivity {
 
         mVerticalSampleButton = (Button) findViewById(R.id.activity_main_vertical_linear_sample_button);
         mVerticalSampleButton.setOnClickListener(mVerticalSampleButtonClickListener);
-
-        mHorizontalSampleButton = (Button) findViewById(R.id.activity_main_horizontal_linear_sample_button);
-        mHorizontalSampleButton.setOnClickListener(mHorizontalSampleButtonClickListener);
-
     }
 
     @NonNull
@@ -43,14 +33,6 @@ public class MainActivity extends AppCompatActivity {
         @Override
         public void onClick(@NonNull View v) {
             startActivity(VerticalLinearRecyclerViewSampleActivity.newIntent(v.getContext()));
-        }
-    };
-
-    @NonNull
-    private View.OnClickListener mHorizontalSampleButtonClickListener = new View.OnClickListener() {
-        @Override
-        public void onClick(@NonNull View v) {
-            startActivity(HorizontalLinearRecyclerViewSampleActivity.newIntent(v.getContext()));
         }
     };
 }