+++ sampleapp/src/main/java/com/bignerdranch/expandablerecyclerviewsample/linear/vertical/Ingredient.java
@@ -1,9 +1,17 @@
 package com.bignerdranch.expandablerecyclerviewsample.linear.vertical;
 
-public class Ingredient {
+import io.realm.RealmObject;
+import io.realm.annotations.PrimaryKey;
 
+public class Ingredient extends RealmObject {
+
+    @PrimaryKey
     private String mName;
     private boolean mIsVegetarian;
+
+    public Ingredient() {
+        // realm constructor
+    }
 
     public Ingredient(String name, boolean isVegetarian) {
         mName = name;