+++ sampleapp/src/main/java/com/bignerdranch/expandablerecyclerviewsample/linear/vertical/VerticalLinearRecyclerViewSampleActivity.java
@@ -10,11 +10,10 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.widget.Toast;
 
-import com.bignerdranch.expandablerecyclerview.ExpandableRecyclerAdapter;
+import com.bignerdranch.expandablerecyclerview.RealmExpandableRecyclerAdapter;
 import com.bignerdranch.expandablerecyclerviewsample.R;
 
-import java.util.Arrays;
-import java.util.List;
+import io.realm.Realm;
 
 /**
  * Sample Activity for the vertical linear RecyclerView sample.
@@ -25,6 +24,8 @@ import java.util.List;
  * @since 5/27/2015
  */
 public class VerticalLinearRecyclerViewSampleActivity extends AppCompatActivity{
+
+    private Realm realm;
 
     private RecipeAdapter mAdapter;
 
@@ -38,25 +39,39 @@ public class VerticalLinearRecyclerViewSampleActivity extends AppCompatActivity{
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_recycler_view_sample);
 
-        Ingredient beef = new Ingredient("beef", false);
-        Ingredient cheese = new Ingredient("cheese", true);
-        Ingredient salsa = new Ingredient("salsa", true);
-        Ingredient tortilla = new Ingredient("tortilla", true);
-        Ingredient ketchup = new Ingredient("ketchup", true);
-        Ingredient bun = new Ingredient("bun", true);
+        realm = Realm.getDefaultInstance();
 
-        Recipe taco = new Recipe("taco", Arrays.asList(beef, cheese, salsa, tortilla));
-        Recipe quesadilla = new Recipe("quesadilla", Arrays.asList(cheese, tortilla));
-        Recipe burger = new Recipe("burger", Arrays.asList(beef, cheese, ketchup, bun));
-        final List<Recipe> recipes = Arrays.asList(taco, quesadilla, burger);
+//        Ingredient beef = new Ingredient("beef", false);
+//        Ingredient cheese = new Ingredient("cheese", true);
+//        Ingredient salsa = new Ingredient("salsa", true);
+//        Ingredient tortilla = new Ingredient("tortilla", true);
+//        Ingredient ketchup = new Ingredient("ketchup", true);
+//        Ingredient bun = new Ingredient("bun", true);
+//
+//        RealmList<Ingredient> tacoIngredients = new RealmList<>();
+//        tacoIngredients.addAll(Arrays.asList(beef, cheese, salsa, tortilla));
+//        Recipe taco = new Recipe("taco", tacoIngredients);
+//
+//        RealmList<Ingredient> quesadillaIngredients = new RealmList<>();
+//        quesadillaIngredients.addAll(Arrays.asList(cheese, tortilla));
+//        Recipe quesadilla = new Recipe("quesadilla", quesadillaIngredients);
+//
+//        RealmList<Ingredient> burgerIngredients = new RealmList<>();
+//        burgerIngredients.addAll(Arrays.asList(beef, cheese, ketchup, bun));
+//        Recipe burger = new Recipe("burger", burgerIngredients);
+//
+//        final List<Recipe> recipes = Arrays.asList(taco, quesadilla, burger);
+//        realm.beginTransaction();
+//        realm.insertOrUpdate(recipes);
+//        realm.commitTransaction();
 
         RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recyclerview);
-        mAdapter = new RecipeAdapter(this, recipes);
-        mAdapter.setExpandCollapseListener(new ExpandableRecyclerAdapter.ExpandCollapseListener() {
+        mAdapter = new RecipeAdapter(this, realm.where(Recipe.class).findAll());
+        mAdapter.setExpandCollapseListener(new RealmExpandableRecyclerAdapter.ExpandCollapseListener() {
             @UiThread
             @Override
             public void onParentExpanded(int parentPosition) {
-                Recipe expandedRecipe = recipes.get(parentPosition);
+                Recipe expandedRecipe = mAdapter.getItem(parentPosition);
 
                 String toastMsg = getResources().getString(R.string.expanded, expandedRecipe.getName());
                 Toast.makeText(VerticalLinearRecyclerViewSampleActivity.this,
@@ -68,7 +83,7 @@ public class VerticalLinearRecyclerViewSampleActivity extends AppCompatActivity{
             @UiThread
             @Override
             public void onParentCollapsed(int parentPosition) {
-                Recipe collapsedRecipe = recipes.get(parentPosition);
+                Recipe collapsedRecipe = mAdapter.getItem(parentPosition);
 
                 String toastMsg = getResources().getString(R.string.collapsed, collapsedRecipe.getName());
                 Toast.makeText(VerticalLinearRecyclerViewSampleActivity.this,
@@ -92,5 +107,11 @@ public class VerticalLinearRecyclerViewSampleActivity extends AppCompatActivity{
     protected void onRestoreInstanceState(@NonNull Bundle savedInstanceState) {
         super.onRestoreInstanceState(savedInstanceState);
         mAdapter.onRestoreInstanceState(savedInstanceState);
+    }
+
+    @Override
+    protected void onDestroy() {
+        realm.close();
+        super.onDestroy();
     }
 }