+++ CHANGELOG.md
@@ -20,8 +20,8 @@ Version 3.0.0-RC1 (10/07/16)
 - Added support for multiple view types within recyclerview
     - `getParentViewType()`, `getChildViewType()`, and `isParentViewType()` added with default implementation for single view type
     - Modified signatures of `onCreateParentViewHolder()` and `onCreateChildViewHolder()` to pass view type being created
-- `ParentViewHolder` package location modified from `com.bignerdranch.expandablerecyclerview.Adapter.ExpandableRecyclerAdapter` to `com.bignerdranch.expandablerecyclerview.ExpandableRecyclerAdapter`
-- `ChildViewHolder` package location modified from `com.bignerdranch.expandablerecyclerview.Adapter.ExpandableRecyclerAdapter` to `com.bignerdranch.expandablerecyclerview.ExpandableRecyclerAdapter`
+- `ParentViewHolder` package location modified from `com.bignerdranch.expandablerecyclerview.Adapter.ExpandableRecyclerAdapter` to `com.bignerdranch.expandablerecyclerview.RealmExpandableRecyclerAdapter`
+- `ChildViewHolder` package location modified from `com.bignerdranch.expandablerecyclerview.Adapter.ExpandableRecyclerAdapter` to `com.bignerdranch.expandablerecyclerview.RealmExpandableRecyclerAdapter`
 - Added ParentViewHolder methods:
     - `getParent()`
     - `getParentAdapterPosition()`